Moved the percent macro lines to the top of the file where people look for them as they were buried a long way down the file, making sure to attach the comment line block to the first record in the file (no blank line).
Also moved the percent doc file description onto one line, as tools do not
pick up following lines.

I have not changed the order of the records or changed any other text, I used
a text editor.

The problem is that for VDCT, comments have to be attached to a record, it
does not support file level comments and also that VDCT re-orders the records
when one is changed, which happened in 2011.  This was actually fixed in 2012
as part of some other work, which was then subsequently backed out, along with
the fix for this.